Add test when simple view is interpreted as list view
Thing is, spectacular uses 'is_list_view'
to determine whether view is listing or not.
If it is then response is wrapped in array
and operation ID has suffix '_list'.

'is_list_view' has some heuristics,
and the last of them is to explode route
and check if the last part is dynamic parameter like {id}.

So, if you have APIView with route
that doesn't end with such parameter,
you will end up with response serializer wrapped with array
just because of the failed heuristic!