Move system check registration to AppConfig
Registering the system check in the __init__.py file has the unfortunate
side effect of registering the check even when the app is not installed.
This can be problematic as DRF settings may also be missing and
`DEFAULT_SCHEMA_CLASS` may not be setup correctly. This will result in
erroneous failed system checks that cannot be solved easily.