aid new users to a more sane default.
inclusion of the schema into the schema is a misguided feature
and simply a unnessary boasting of capabilites. Can't turn it
of by default anymore as it may influence operation naming and
thus break users. But new users may copy&paste this snippet
and have less issues with automatic prefix estimation.