Avoid using default role in documentation.
Sphinx's default role (wrapping in single backticks) is often confusing
because it doesn't render as code - which achieved using double
backticks in reStructuredText - and this differs from Markdown. In
Sphinx, the default role is rendered in the same way as wrapping with
single asterisks by default - providing italicised emphasis.

To avoid this confusion, we can override the default role to warn us
when we are using it and change existing usages to use correct markup -
double backticks where we want code, single asterisks where we want
emphasis.